Beta version of CSRF protection

+++ src/backend/controllers/purchasesController.js
@@ -1,15 +1,18 @@
+import session from "express-session";
 import pool from "../db/index.js";
 
 class PurchasesController {
   async getPurchases(req, res) {
-    const  id  = req.params.id; 
+    const id = req.params.id;
     try {
       const purchases = await pool.query(
         `SELECT * FROM purchases_channels WHERE user_id = $1`,
         [id]
       );
-      if(purchases.rows.length === 0){
-        res.status(404).json({message:"Не удалось получить покупки пользователя"})
+      if (purchases.rows.length === 0) {
+        res
+          .status(404)
+          .json({ message: "Не удалось получить покупки пользователя" });
       }
       res.json(purchases.rows[0]);
     } catch (error) {
@@ -20,42 +23,52 @@ class PurchasesController {
   async addPurchase(req, res) {
     const { id } = req.params; // Получаем id пользователя из параметров
     const { uses, thumbnail, email, channelName } = req.body; // Получаем данные из тела запроса
-  
+
+    const tokenFromClient = req.headers["x-csrf-token"];
+    const tokenFromSession = req.session.csrfToken;
+
+    if (tokenFromClient !== tokenFromSession) {
+      return res.status(403).send('CSRF token mismatch');
+    }
+
     try {
       // Проверяем, существует ли пользователь
-      const userCheck = await pool.query(`SELECT * FROM users WHERE user_id = $1`, [id]);
+      const userCheck = await pool.query(
+        `SELECT * FROM users WHERE user_id = $1`,
+        [id]
+      );
       if (userCheck.rows.length === 0) {
         return res.status(404).json({ message: "Пользователь не найден" });
       }
-  
+
       // Проверяем, существует ли покупка для указанного канала
       const channelsCheck = await pool.query(
         `SELECT * FROM purchases_channels WHERE channel_name = $1 AND user_id = $2`,
         [channelName, id]
       );
-  
+
       if (channelsCheck.rows.length > 0) {
         // Если канал уже куплен, возвращаем соответствующий ответ
         return res
           .status(409)
           .json({ message: "У вас уже приобретены данные этого канала!" });
       }
-  
+
       const user = userCheck.rows[0];
-  
+
       // Проверяем, достаточно ли использований на балансе
       if (user.uses < uses) {
         return res
           .status(402)
           .json({ message: "Недостаточно использований на балансе!" });
       }
-  
+
       // Обновляем баланс использований
       const updateUses = await pool.query(
         `UPDATE users SET uses = uses - $1 WHERE user_id = $2 RETURNING *`,
         [uses, id]
       );
-  
+
       // Добавляем покупку в таблицу purchases_channels
       const purchase = await pool.query(
         `INSERT INTO purchases_channels (user_id, channel_name, thumbnail, email) 
@@ -63,7 +76,7 @@ class PurchasesController {
          RETURNING thumbnail, email, channel_name`,
         [id, channelName, thumbnail, email]
       );
-  
+
       // Возвращаем результат клиенту
       res.json({
         message: "Вы успешно приобрели данные ютубера",
@@ -75,20 +88,28 @@ class PurchasesController {
       res.status(500).json({ message: "Ошибка сервера", error: error.message });
     }
   }
-  
-  
 
   async deletePurchase(req, res) {
-    const {channelName} = req.body;
-    const id = req.params.id
-    console.log("Id и channelName:",id,channelName)
+    const { channelName } = req.body;
+    const id = req.params.id;
+
+    const tokenFromClient = req.headers["x-csrf-token"];
+    const tokenFromSession = req.session.csrfToken;
+
+    if (tokenFromClient !== tokenFromSession) {
+      return res.status(403).send('CSRF token mismatch');
+    }
+    
+    console.log("Id и channelName:", id, channelName);
     try {
       const deleteOperation = await pool.query(
         `DELETE FROM purchases_channels WHERE user_id = $1 AND channel_name = $2 RETURNING *`,
         [id, channelName]
       );
-      if(deleteOperation.rows.length === 0){
-        res.status(400).json({message:"Возникла ошибка в удалении покупки."})
+      if (deleteOperation.rows.length === 0) {
+        res
+          .status(400)
+          .json({ message: "Возникла ошибка в удалении покупки." });
       }
       res.json(deleteOperation.rows[0]);
     } catch (error) {
